server-side/pyproject.toml
[project]
name = "scrape-site"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=10.4.0",
    "playwright>=1.47.0",
    "werkzeug>=3.0.4",
]
